Block loop flow

+++ meddb/db.go
@@ -24,7 +24,7 @@ type BlockchainDB interface {
 	// Writes vote to vote table
 	WriteVote(*Vote) error
 	// Returns all votes for given public key from votes table with the given VotedAt
-	GetVotes([]byte, int) ([]*Vote, error)
+	GetVotes([]byte, int64) ([]*Vote, error)
 	// Returns k most recent votes for given public key from votes table sorted by decreasing
 	// VotedAt timestamp.
 	GetRecentVotes([]byte, int) ([]*Vote, error)
